설비 1차 작업 완료 (추후 피드백 들어오는거 수정하면 될듯)

+++ main/java/kr/or/chop/P16_equip/dto/EqDTO.java
@@ -27,8 +27,8 @@ public class EqDTO {
 	
 	
 	// 설비 통계 (임시 계산값)
-	private int totalTime;        // 전체시간
-	private int runTime;          // 가동시간
+	private double totalTime;        // 전체시간
+	private double runTime;          // 가동시간
 	private String runRate;       // 가동률
